Jwt RefreshToken added

+++ src/main/java/com/conner/assistant/repository/RefreshTokenRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Integer> {
     Optional<RefreshToken> findByToken(String token);
 }
